extract grpc into separated configuration
